refactor: Fix typo in import statement for AuthController
