+++ src/routes/authRouter.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-import { AuthController } from "../controllers/AuthContoller";
+import { AuthController } from "../controllers/AuthController";
 import { body, param } from "express-validator";
 import { handleInputErrors } from "../middleware/validation";
 import { limiter } from "../config/limiter";
